+++ substrate/runtime/council/src/lib.rs
@@ -468,7 +468,7 @@ impl<T: Trait> Module<T> {
 			<NextFinalise<T>>::put((number + Self::presentation_duration(), empty_seats as u32, expiring));
 
 			let voters = Self::voters();
-			let votes = voters.iter().map(<staking::Module<T>>::balance).collect::<Vec<_>>();
+			let votes = voters.iter().map(<staking::Module<T>>::voting_balance).collect::<Vec<_>>();
 			<SnapshotedStakes<T>>::put(votes);
 
 			// initialise leaderboard.
